Refactor, feat, add: 사운드 추가, 공부모드 1시간 진행시, 엑션취하라는 모달추가, 폴더구조 변경
