+++ src/util/images/index.js
@@ -1,9 +1,9 @@
-import turtleneck from './turtleneck.gif';
-// import arm from './arm.gif';
-// import sideneck from './sideneck.gif';
+import turtleNeck from './turtleneck_stretching.gif';
+import sideArmStretching from './side_arm_stretching.gif';
+import sideNeck from './side_neck_stretching.gif';
 
 export const poseImage = {
-  TurtleNeck: turtleneck,
-  // Arm: arm,
-  // SideNeck: sideneck,
+  TurtleNeck: turtleNeck,
+  Arm: sideArmStretching,
+  SideNeck: sideNeck,
 };
