+++ src/util/music/index.js
@@ -1,7 +1,9 @@
 import okGood from './ok_sound.mp3';
 import timeShutdown from './timer_shutdown.mp3';
 import warningSound from './warning_sound.mp3';
+import checkSound from './check_alarm.mp3';
 
 export { okGood };
 export { warningSound };
 export { timeShutdown };
+export { checkSound };
